refactor(api.rs,app.rs,home/mod.rs): Moved the authentication check api call to the home component
